style: run ktfmtFormat

+++ src/main/kotlin/model/algorithms/Djikstra.kt
@@ -3,7 +3,7 @@ package model.algorithms
 import model.graph.Graph
 
 class Djikstra(private val graph: Graph, private val startVertexID: Int = -1) {
-    private val distance = hashMapOf<Int, Long>()
+    private val distance = hashMapOf<Int, Float>()
     private val visited = hashMapOf<Int, Boolean>()
     private val from = hashMapOf<Int, Int>()
     private val n = graph.vertices.size
@@ -12,32 +12,33 @@ class Djikstra(private val graph: Graph, private val startVertexID: Int = -1) {
         if (n == 0 || startVertexID <= -1) return
 
         for ((id, _) in graph.vertices) {
-            distance[id] = Long.MAX_VALUE
+            distance[id] = Float.MAX_VALUE
             from[id] = -1
         }
 
-        distance[startVertexID] = 0
+        distance[startVertexID] = 0f
 
         for (i in 0 until n) {
 
             var nearest = -1
             for ((vertexID, _) in graph.vertices) {
-                if (!visited.getOrDefault(
-                        vertexID,
-                        false
-                    ) && (nearest == -1 || distance[vertexID]!! < distance[nearest]!!)
+                if (
+                    !visited.getOrDefault(vertexID, false) &&
+                        (nearest == -1 || distance[vertexID]!! < distance[nearest]!!)
                 ) {
                     nearest = vertexID
                 }
             }
             visited[nearest] = true
 
-            if (distance[nearest] == Long.MAX_VALUE) break
+            if (distance[nearest] == Float.MAX_VALUE) break
 
             for (edgeID in graph.vertices[nearest]!!.incidentEdges) {
                 val edge = graph.edges[edgeID]!!
 
-                val to = if (nearest == edge.vertices.first) edge.vertices.second else edge.vertices.first
+                val to =
+                    if (nearest == edge.vertices.first) edge.vertices.second
+                    else edge.vertices.first
                 val weight = edge.weight
 
                 if (distance[nearest]!! + weight < distance[to]!!) {
