+++ src/main/kotlin/model/databases/CSV/data/CSVGraphData.kt
@@ -1,7 +1,6 @@
 package model.databases.CSV.data
 
 data class CSVGraphData(
-
     var isNode: Boolean,
     var id: Int,
     var name: String,
@@ -10,8 +9,7 @@ data class CSVGraphData(
     var color: String?,
     var radius: Double?,
     var community: Int?,
-
     var from: String?,
     var to: String?,
-    var weight: Long?
+    var weight: Float?
 )
