+++ src/main/kotlin/model/databases/sqlite/dao/edge/Edges.kt
@@ -6,5 +6,5 @@ import org.jetbrains.exposed.dao.id.IntIdTable
 object Edges : IntIdTable("Edges") {
     val first = reference("first", Vertices).nullable()
     val second = reference("second", Vertices).nullable()
-    val weight = long("weight")
+    val weight = float("weight")
 }
