+++ src/main/kotlin/model/graph/Edge.kt
@@ -2,8 +2,8 @@ package model.graph
 
 class Edge(
     val vertices: Pair<Int, Int>,
-    var weight: Long = 1L,
+    var weight: Float = 1f,
     var id: Int = 1,
 ) {
-    //fun incident(v: Int) = v == vertices.first || v == vertices.second
+    // fun incident(v: Int) = v == vertices.first || v == vertices.second
 }
